Revert "Merge pull request #2330 from berryzplus/feature/apply_wall"
This reverts commit 50e86a4ca56ef49018761fd2a2d7be61e9922a7a, reversing
changes made to c3c24ce9491dabfd89378f4afb515a11a5afa4d4.

+++ sakura_core/_main/CControlTray.cpp
@@ -987,7 +987,7 @@ void CControlTray::OnCommand( WORD wNotifyCode, [[maybe_unused]] WORD wID , [[ma
 
 bool CControlTray::OnSetTypeSetting(size_t index)
 {
-	if (m_pShareData->m_nTypesCount <= 0 || size_t(m_pShareData->m_nTypesCount) <= index) {
+	if (m_pShareData->m_nTypesCount <= 0 || m_pShareData->m_nTypesCount <= index) {
 		return false;
 	}
 
@@ -1012,7 +1012,7 @@ bool CControlTray::OnSetTypeSetting(size_t index)
 
 bool CControlTray::OnGetTypeSetting(size_t index)
 {
-	if (m_pShareData->m_nTypesCount <= 0 || size_t(m_pShareData->m_nTypesCount) <= index) {
+	if (m_pShareData->m_nTypesCount <= 0 || m_pShareData->m_nTypesCount <= index) {
 		return false;
 	}
 
@@ -1023,7 +1023,7 @@ bool CControlTray::OnGetTypeSetting(size_t index)
 
 bool CControlTray::OnAddTypeSetting(size_t index)
 {
-	if (m_pShareData->m_nTypesCount < 0 || int(MAX_TYPES) <= m_pShareData->m_nTypesCount || size_t(m_pShareData->m_nTypesCount) < index) {
+	if (m_pShareData->m_nTypesCount < 0 || int(MAX_TYPES) <= m_pShareData->m_nTypesCount || m_pShareData->m_nTypesCount < index) {
 		return false;
 	}
 
@@ -1066,7 +1066,7 @@ bool CControlTray::OnAddTypeSetting(size_t index)
 
 bool CControlTray::OnDelTypeSetting(size_t index)
 {
-	if (m_pShareData->m_nTypesCount <= 0 || size_t(m_pShareData->m_nTypesCount) <= index) {
+	if (m_pShareData->m_nTypesCount <= 0 || m_pShareData->m_nTypesCount <= index) {
 		return false;
 	}
